Fixed checkstyle warnings
